pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mars-llm"
version = "0.0.3"
description = "Multi-Agent Resolution Synthesis — LLMs debate to find the best answer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Jeremy Pollock" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "openai>=1.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "google-auth>=2.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-slugify>=8.0",
]

[project.scripts]
mars = "mars.cli:main"

[project.urls]
Homepage = "https://github.com/jpollock/multi-agent-resolution-synthesis"
Repository = "https://github.com/jpollock/multi-agent-resolution-synthesis"

[tool.hatch.build.targets.wheel]
packages = ["src/mars"]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
mypy_path = "src"
packages = ["mars"]
